requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
groq
python-dotenv
